+++ packages/yoshi-flow-editor/src/commands/build.ts
@@ -79,6 +79,7 @@ const build: cliCommand = async function(argv, config, model) {
 
     const petriSpecs = await import('yoshi-common/build/sync-petri-specs');
     const wixMavenStatics = await import('yoshi-common/build/maven-statics');
+    const copyDocker = await import('yoshi-common/build/copy-docker');
 
     await Promise.all([
       petriSpecs.default({
@@ -88,6 +89,7 @@ const build: cliCommand = async function(argv, config, model) {
         clientProjectName: config.clientProjectName,
         staticsDir: config.clientFilesPath,
       }),
+      copyDocker.default(config),
     ]);
   }