+++ test/typescript/features/yoshi-server/api-error-message/src/api/greeting.api.ts
@@ -1,0 +1,11 @@
+import { method } from 'yoshi-server';
+
+export const greet = method(function(name: string) {
+  if (name === 'Yaniv') {
+    throw new Error('A cool error!');
+  }
+  return {
+    greeting: `hello ${name}`,
+    name,
+  };
+});